+++ Configuration/JavaScriptModules.php
@@ -1,0 +1,11 @@
+<?php
+
+return [
+    'dependencies' => ['core', 'backend'],
+    'tags' => [
+        'backend.contextmenu',
+    ],
+    'imports' => [
+        '@t3docs/examples/' => 'EXT:examples/Resources/Public/JavaScript/',
+    ],
+];